# openapi_sveta.yaml
openapi: 3.0.3
info:
  title: API spb4you 3.0
  version: 1.0.11
  description: Все необходимые API для сайта Питер_Для_Тебя
...
paths:
...
  /locations/info/{locationId}:
    get:
      tags:
        - LocationsController
      summary: Получение информации об определенной локации
      description: Возвращает набор полей локации по соответствующему id
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешный ответ со списком полей определенной локации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
  /locations/{locationId}:
    get:
      tags:
        - LocationsController
      summary: Получение страницы определенной локации
      description: Возвращает страницу локации по соответствующему id
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешный ответ и вывод страницы локации
...
  /locations/info/listbytag/{tagId}:
    get:
      tags:
        - LocationsController
      summary: Получение списка локаций с определенным тегом
      description: Возвращает список всех локаций с определенным тегом
      parameters:
        - name: tagId
          in: path
          required: true
          description: Уникальный идентификатор тега
          schema:
            type: integer
      responses:
        '200':
          description: Успешное получение списка локаций
        '404':
          description: Тег не найден
...
  /routes/info/{routeId}:
    get:
      tags:
        - RoutesController
      summary: Получение информации об определенном маршруте
      description: Возвращает набор полей маршрута по соответствующему id
      parameters:
        - name: routeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешный ответ со списком полей определенной локации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
  /routes/{routeId}:
    get:
      tags:
        - RoutesController
      summary: Получение страницы определенного маршрута
      description: Возвращает страницу маршрута по соответствующему id
      parameters:
        - name: routeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешный ответ и вывод страницы маршрута
...
  /routes/info/listbytag/{tagId}:
    get:
      tags:
        - RoutesController
      summary: Получение списка маршрутов с определенным тегом
      description: Возвращает список всех маршрутов с определенным тегом
      parameters:
        - name: tagId
          in: path
          required: true
          description: Уникальный идентификатор тега
          schema:
            type: integer
      responses:
        '200':
          description: Успешное получение списка маршрутов
        '404':
          description: Тег не найден
...
  /tags/info/{tagId}:
    get:
      tags:
        - TagsController
      summary: Получение информации об определенном теге
      description: Возвращает набор полей тега по соответствующему id
      parameters:
        - name: tagId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешный ответ со списком полей определенного тега
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
...
  /tags/{tagId}:
    get:
      tags:
        - TagsController
      summary: Получение страницы с локациями по тегу
      description: Возвращает страницу со списком локаций по id определенного тега
      parameters:
        - name: tagId
          in: path
          required: true
          description: Уникальный идентификатор тега
          schema:
            type: integer
      responses:
        '200':
          description: Успешный ответ и вывод страницы с локациями по тегу
...
  /categories/info/{categoryId}:
    get:
      tags:
        - CategoriesController
      summary: Получение информации об определенной категории
      description: Возвращает набор полей категории по соответствующему id
      parameters:
        - name: categoryId
          in: path
          required: true
          description: Уникальный идентификатор категории
          schema:
            type: integer
      responses:
        '200':
          description: Успешный ответ со списком полей определенной категории
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /categories/locations:
    get:
      tags:
        - CategoriesController
      summary: Получить список категорий для выбора локаций
      description: Возвращает страницу с категориями, тип которых 'Локации'
      responses:
        '200':
          description: Страница с категориями для выбора локаций
        '500':
          description: Ошибка сервера
  /categories/locations/{categoryId}:
    get:
      tags:
        - CategoriesController
      summary: Получить список локаций по категории
      description: Возвращает страницу с локациями, которые принадлежат указанной категории по id категории
      parameters:
        - name: categoryId
          in: path
          required: true
          description: Уникальный идентификатор категории
          schema:
            type: integer
      responses:
        '200':
          description: Страница с локациями для указанной категории
        '404':
          description: Категория не найдена
        '500':
          description: Ошибка сервера
  /categories/routes:
    get:
      tags:
        - CategoriesController
      summary: Получить список категорий для выбора маршрутов
      description: Возвращает страницу с категориями, тип которых 'Маршруты'
      responses:
        '200':
          description: Страница с категориями для выбора маршрутов
        '500':
          description: Ошибка сервера
  /categories/routes/{categoryId}:
    get:
      tags:
        - CategoriesController
      summary: Получить список маршрутов по категории
      description: Возвращает страницу с маршрутами, которые принадлежат указанной категории по id категории
      parameters:
        - name: categoryId
          in: path
          required: true
          description: ID категории, для которой необходимо получить маршруты
          schema:
            type: integer
      responses:
        '200':
          description: Страница с маршрутами для указанной категории
        '404':
          description: Категория не найдена
        '500':
          description: Ошибка сервера
components:
  schemas:
    Location:  # Определение сущности "Локация"
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор локации
          example: 4
        name:
          type: string
          description: Название локации
          example: "Лофт проект ЭТАЖИ"
        tagId:
          type: integer
          description: Уникальный идентификатор тега
          example: 1
        mainPhoto:
          type: integer
          description: ID основной фотографии
          example: 135
        photos:
          type: array
          items:
            type: integer
          description: Список ID фотографий
          example: [21, 65, 79]
        description:
          type: string
          description: Описание локации
          example: "Самое популярное место..."
        geoData:
          type: array
          items:
            type: number
          description: Географическая ширина и долгота
          example: [35.8975, 45.4698]
        additInfo:
          type: string
          description: Дополнительная информация
          example: "Адрес: ... Часы работы: ..."
        likes:
          type: integer
          description: Количество лайков
          example: 125
      required:
        - id
        - name
        - tagId
        - mainPhoto
        - photos
        - description
        - geoData
        - additInfo
        - likes
...
    Tag: # Определение сущности "Тег"
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор тега
          example: 1
        name:
          type: string
          description: Название тега
          example: "Дворы-рестораны"
      
      required:
        - id
        - name
    Category: # Определение сущности "Категория"
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор категории
          example: 1
        name:
          type: string
          description: Название категории
          example: "Для фото"
        description:
          type: string
          description: Дополнительное описание категории
          example: В данной подборке вы найдете для себя места, где можно сделать красивые фотографии. Различные локации на улицах и во дворах Петербурга смогут разбавить вашу ленту фотографий.
        type:
          type: string
          description: Тип категории
          example: "Локации"  
        second_name:
          type: string
          description: Второе название категории
          example: "Локации для фото"
      required:
        - id
        - name
...